mkcloud: Simplify the lonely/ironic nodes setup with setupallnodes
Instead of the need to call the steps "setuplonelynodes" and
"setupironicnodes" manually, do that whenever "setupnodes" was
called. This is done in a new step called "setupallnodes" which
includes "setupnodes".
If the variables to handle the node numbers (nodenumberlonelynode and
nodenumberironicnode) are 0 (the default) nothing happens.

This simplifies the setup of theses node types. You just need to
export the nodenumber* variables and run the usual mkcloud steps (like
plain, all_noreboot, ...) to get lonely/ironic nodes included.